requirements.txt
style-bert-vits2>=0.1.0
fastapi==0.115.2
uvicorn==0.23.2
python-dotenv==1.0.0
google-generativeai==0.3.1
pydub==0.25.1
requests==2.31.0
numpy==1.26.0
simpleaudio==1.0.4
gradio==5.20.1